Changed statistics from single to double precision

+++ Application/Application.h
@@ -21,8 +21,8 @@ namespace NetworkLoad {
         void syncRanksAndPrint();
         void syncConfiguration();
 
-        float startMessaging();
-        float gatherConfidentMessage();
+        double startMessaging();
+        double gatherConfidentMessage();
 
     private:
         Configuration configuration;
